API/requirements.txt
fastapi
uvicorn[standard]
ultralytics
torch
torchvision
pdf2image
pytesseract
opencv-python-headless
Pillow
python-multipart
